Improve file creation with paths
Remove URI repository due it can be accessed by URI and LocalPath. As
repository key is URI, when we acces by LocalPath it doesn't have sense

+++ src/megaapi_impl.cpp
@@ -3533,7 +3533,7 @@ bool MegaTransferPrivate::serialize(string *d) const
     }
 
     auto localPathSerialized = mLocalPath.serialize();
-    ll = localPathSerialized.size();
+    ll = static_cast<unsigned short>(localPathSerialized.size());
     d->append((char*)&ll, sizeof(ll));
     d->append(localPathSerialized.data(), ll);
